src/customer_insights/data/requirements.txt
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
